+++ classes/test.php
@@ -178,7 +178,7 @@ abstract class test implements observable, \countable
 
 				$annotationExtractor->extract($publicMethod->getDocComment());
 
-				if ($publicMethod->getNumberOfParameters() > 0 && isset($this->dataProviders[$methodName]) === false)
+				if ($this->methodIsIgnored($methodName) === false && $publicMethod->getNumberOfParameters() > 0 && isset($this->dataProviders[$methodName]) === false)
 				{
 					$this->setDataProvider($methodName);
 				}
@@ -1266,7 +1266,16 @@ abstract class test implements observable, \countable
 					}
 					else
 					{
-						$data = $this->{$this->dataProviders[$testMethod]}();
+						$dataProvider = $this->dataProviders[$testMethod];
+
+						if ($dataProvider instanceof test\data\provider)
+						{
+							$data = $this->dataProviders[$testMethod]();
+						}
+						else
+						{
+							$data = $this->{$this->dataProviders[$testMethod]}();
+						}
 
 						if (is_array($data) === false && $data instanceof \traversable === false)
 						{
@@ -1283,9 +1292,9 @@ abstract class test implements observable, \countable
 								$arguments = array($arguments);
 							}
 
-							if (sizeof($arguments) != $numberOfArguments)
+							if (sizeof($arguments) < $numberOfArguments)
 							{
-								throw new test\exceptions\runtime('Data provider ' . $this->getClass() . '::' . $this->dataProviders[$testMethod] . '() not provide enough arguments at key ' . $key . ' for test method ' . $this->getClass() . '::' . $testMethod . '()');
+								throw new test\exceptions\runtime('Data provider ' . ($dataProvider instanceof test\data\provider ? '' : $this->getClass() . '::' . $this->dataProviders[$testMethod] . '() ') . 'does not provide enough arguments at key ' . $key . ' for test method ' . $this->getClass() . '::' . $testMethod . '()');
 							}
 
 							$this->score->setDataSet($key, $this->dataProviders[$testMethod]);
@@ -1431,9 +1440,34 @@ abstract class test implements observable, \countable
 		if ($dataProvider === null)
 		{
 			$dataProvider = $testMethodName . 'DataProvider';
+
+			if (method_exists($this->checkMethod($testMethodName), $dataProvider) === false)
+			{
+				$reflectedMethod = call_user_func($this->reflectionMethodFactory, $this, $testMethodName);
+				$parametersProvider = new test\data\provider\aggregator();
+
+				foreach ($reflectedMethod->getParameters() as $parameter)
+				{
+					$parameterProvider = new test\data\providers\mock($this->mockGenerator);
+
+					if (($parameterClass = $parameter->getClass()) === null)
+					{
+						throw new exceptions\logic\invalidArgument('Could not generate a data provider for ' . $this->class . '::' . $testMethodName . '() because it has at least one argument which is not type-hinted with a class or interface name');
+					}
+
+					$parametersProvider->addProvider($parameterProvider->setClass($parameterClass->getName()));
+				}
+
+				$dataProvider = new test\data\set($parametersProvider);
+			}
 		}
 
-		if (method_exists($this->checkMethod($testMethodName), $dataProvider) === false)
+		if ($dataProvider instanceof \closure)
+		{
+			throw new exceptions\logic\invalidArgument('Cannot use a closure as a data provider for method ' . $this->class . '::' . $testMethodName . '()');
+		}
+
+		if ($dataProvider instanceof test\data\provider === false && method_exists($this->checkMethod($testMethodName), $dataProvider) === false)
 		{
 			throw new exceptions\logic\invalidArgument('Data provider ' . $this->class . '::' . lcfirst($dataProvider) . '() is unknown');
 		}