+++ classes/test/data/provider.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace mageekguy\atoum\test\data;
+
+
+interface provider
+{
+	public function __invoke();
+
+	public function __toString();
+
+	public function generate();
+}