+++ classes/test/data/provider/object/exceptions/mandatoryArgument.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace mageekguy\atoum\test\data\provider\object\exceptions;
+
+use mageekguy\atoum;
+
+
+class mandatoryArgument extends atoum\exceptions\runtime {}